Cache parse results
This eliminates the exponential growth of tree traversal, since once we
decend one time, we don't need to descend again.

This solution overdoes the caching, though. If a production is only
derivable from one possible parent, there's no need to cache it. (The
parent will do that for us.) Cleaning that up is a topic for the future.

(In the meantime: we can see it in the coverage reports. Many of the
productions will never have their cached value read, and so the code
that supports that never runs. To be fair, the tests aren't complete
enough to check the ones that _could_ be read, either.)

Fixes #1.