Remove string interning
I tried switching to a [u16] interner instead (which worked fine), but I
was having lifetime problems dealing with the underlying buffers when it
came time to _use_ them. The impact is that I'd have to _copy_ out of
the pool every time we wanted to use a string, and if that's the case,
what's the point?

I'll come back to this, maybe, in the future. It doesn't really impact
this as a toy project, after all.