Switch GitHub actions to nightly; add clippy and format checks (#65)
The latest changes I threw in to fix some clippy errors in the tests require
an unstable interface, so are only part of nightly compiler toolchains, not
the stable toolchains. 

This changes installs a more sophisticated set of toolchain automation for 
rust, which now (in addition to running the test suite) will also do all those 
clippy checks, and make sure the formatting is good as well. 

The formatting check is especially nice, as I've included a few changes in the
past that are unrelated to the PR in question, and are just formatting fixes
left over from earlier PRs.