Simplify early errors by removing do-nothing forms (#152)
I've been wishy-washy for a while, but finally decided: there's no
reason for callers to be ignorant of the fact that a particular
production doesn't actually have any errors. Or doesn't use it's
"strict" argument. If those things are true, make the calling pattern
cleaner and follow that back up.