Add a `bound_name` api for the productions with only one name
Realized that in a number of places were were taking "the sole element
of `bound_names`". This was a bit silly. Why make a vector of names
only to immediately throw the vector part away? That's a wasted
allocation. So this change lets consumers of the one-name productions
actually just get that one name.