Put agents into Rc<RefCell<>> (effectively)
The needs of a few generators outweigh the needs of the many.

If generators work, it means that we can no longer pass around a single
&mut reference to the agent, as it would be need to be stashed in the
"paused execution" of a generator (and thus generating two mutable
references).

So we pass around a shared reference instead, which means we need to
switch to iterior mutability for the agent.

Also wrapped in an Rc, since the lifetime of any generator object
will outlive anything that creates it, and so passing around the agent
lifetimes as well will otherwise become a bit too onerous.