Add Per-Iteration environment support
And also rename "label set" to "string set", because this is a case
where the code wants a set of strings that aren't labels.

This isn't connected to anything, yet, but hey: I got a complete test
written for it, so that feels like an accomplishment worth celebrating
by turning this into a complete commit.